Added: Class ComponentGuestLabelInstruction
Added this class which allows for specifying of an instruction text at the same location in every window.

NOTE!
This is centered using HMTL inputs to make it stay in place, this also means you have to use the br html tag for line breaking instead of \n for new line.