Added: Class ViewGuestCustomerInformation
Added the class ViewGuestCustomerInformation which is the follow-up window after having filled out the table number in the ViewGuestTableInformation view.

I would also like to point out that the author notes do not correctly reflect the time when these were worked on, as this is partly a result of various work over the course of the last days.