Fixed: Wrong Title
Fixed an issue where the staff window's would use the wrong title from an older project, as the GUI for this part of the system is currently based upon the old project's graphical user interface for the demo.

In future iterations a dedicated graphical user interface structure with its own components should be made for the staff's part of the system as well.