Improved: ViewGuestMenuAdult Layout
Improved on the visual layout of the ViewGuestMenuAdult  view / window so that it would not be full screened due to issue with the content's alignment size etcetera.

Unfortunately to completely fix this, it would require a larger change of the user interface's components, some of the graphical user interface components would have to have their borders reduced, which changes a lot of the layout and changes would need to be made to other components in order to keep the layout looking clean in other areas, so for now this is an improvement but not a fix.

An actual modification of these multiple components would have to be made in a future iteration.