Changed: Main (Code Review)
Performed code review on the class to find and correct potential errors, and enforce the coding standard.

Also made changes to the JavaDocs to reflect changes and inaccuracies in descriptiveness.

Also turned the order of which the windows are launched around, so that the guest window is the one in front.